Resolve multiple instances of a plugin being loaded on plugin reload
Due to the use of IPC to trigger a reload on the other browser
instances, we only need to call "PluginsService.Rebuild" from
the instance initiating the reload. To accomplish this, I've
moved it to its own "rebuildPlugin" method that is called before
the reloadPlugin method in plugin-actions.ts.

We can then skip sending the "injectPlugin" IPC, as it will be
handled by the "reloadPlugin" IPC on the other browser instances.

To stop the backend from injecting multiple plugin instances
in each browser instance, we now pass the entrypoint in the
"InjectService.InjectPlugin" call, instead of looping in that
method.

Because there is now a call for rebuilding the specific plugin
we are reloading, a "service.plugins.Reload()" is no longer
necessary in InjectService.InjectPlugin. However this does
necessitate adding a call to a new reloadPlugins RPC in the
install-plugin instance, so that the new plugin is loaded into
memory, before being injected to all browser instances using
reloadPlugin.

To avoid unnecessary RPC calls, the "setEnabled" calls have
been moved outside of the IPC-triggered load/unload methods,
so that they will only be called once on the triggering
browser instance.